Add codes and test data and report
The report is in the Doc folder, the codes have two parts one is R script which can run the model locally, the other is the shiny app codes including glob, server, ui three parts. The shiny code shows the interaction more intuitively, that is the recommand one. User can do some exploratory with the interface. That is awesome!